requirements.txt
langchain
langchain-community
langchain-openai
unstructured
accelerate
transformers
tokenizers
bitsandbytes
einops
xformers
sentence_transformers
chromadb
hydra-core
openai